requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
pillow>=10.3.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.32.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.232

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Additional requirements for this project ------------------------------------
customtkinter==5.2.0
keyboard==0.13.5
pyserial==3.5
pygame==2.5.2
colorama==0.4.6
numba==0.57.1
pywin32>=223  # Using a more flexible version requirement
mss>=6.1.0  # For screen capture